fixed leaks

+++ ft_printf/libft/ft_strmapi.c
@@ -6,7 +6,7 @@
 /*   By: yizhang <yizhang@student.codam.nl>           +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2022/10/21 16:22:58 by yizhang       #+#    #+#                 */
-/*   Updated: 2022/10/31 09:22:50 by yizhang       ########   odam.nl         */
+/*   Updated: 2023/02/21 10:10:58 by yizhang       ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 char	*ft_strmapi(char const *s, char (*f)(unsigned int, char))
 {
-	size_t	i;
-	size_t	len;
+	int	i;
+	int	len;
 	char	*ret;
 
 	i = 0;
