+++ ft_printf/libft/ft_strnstr.c
@@ -6,13 +6,13 @@
 /*   By: yizhang <yizhang@student.codam.nl>           +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2022/10/05 18:21:57 by yizhang       #+#    #+#                 */
-/*   Updated: 2022/10/31 09:23:31 by yizhang       ########   odam.nl         */
+/*   Updated: 2023/02/21 10:03:10 by yizhang       ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-char	*ft_strnstr(const char *haystack, const char *needle, size_t len)
+char	*ft_strnstr(const char *haystack, const char *needle, int len)
 {
 	char	*hay;
 	char	*n;
